// frontend/mirahelp/src/app/home/interfaces/CorporationRequest.ts
 export interface CorporationRequest {
      name: string | null | undefined;
      category: string | null | undefined;
      description: string | null | undefined;
      street: string | null | undefined;
      numberAdress: string | null | undefined;
      neighborhood: string | null | undefined;
      hourOpening: string | null | undefined;
      hoursClose: string | null | undefined;
      phoneNumber: string | null | undefined;    
      
}